[onert-micro] Fix types for building (#13234)
This pr adds static casts to fix types.

ONE-DCO-1.0-Signed-off-by: Artem Balyshev <a.balyshev@samsung.com>

+++ onert-micro/onert-micro/include/pal/common/PALAveragePool2DCommon.h
@@ -55,9 +55,11 @@ OMStatus AveragePool(const core::Pool2DParams &params, const core::OMRuntimeShap
           // Compute the boundaries of the filter region clamped so as to
           // ensure that the filter window fits in the input array.
           const int filter_x_start = std::max(0, -in_x_origin);
-          const int filter_x_end = std::min(params.filter_w, input_width - in_x_origin);
+          const int filter_x_end = std::min(static_cast<int32_t>(params.filter_w),
+                                            static_cast<int32_t>(input_width - in_x_origin));
           const int filter_y_start = std::max(0, -in_y_origin);
-          const int filter_y_end = std::min(params.filter_h, input_height - in_y_origin);
+          const int filter_y_end = std::min(static_cast<int32_t>(params.filter_h),
+                                            static_cast<int32_t>(input_height - in_y_origin));
 
           float total = 0.f;
           float filter_count = 0;
